requirements.txt
azure-ai-projects
azure-identity
python-dotenv
rich
